updated to fix type error

+++ calc_seed_scripts/get_kpts.py
@@ -39,4 +39,4 @@ def get_kpts_array(atoms,desired_kpra):
             else:
                 kpts = kpts -1
         
-    return kpts
+    return kpts.astype(np.int64)
